Revisit all our kubernetes resources
Due to Covid-19 we are reducing the cost of the all AWS services.

When we lower the resources on the microservices we make sure that services reserve the needed resources.
If they need more resources of course they can reach to the limit and also start autoscaling if needed.

The following changes are proposed after reviewing the Sandbox & Production Grafana.